requirements.txt
# Core dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Optional (for Google Drive upload if needed)
# google-auth-oauthlib>=1.0.0
# google-api-python-client>=2.0.0
